agrega al paquete modelo .deporte, .inscripcion, .marca, .resultado y SistemaJJOO. Comienza con los paquetes de excepciones, servicios y UI

+++ src/modelo/persona/Atleta.java
@@ -3,7 +3,7 @@ package modelo.persona;
 public class Atleta extends Persona {
     private String especialidad;
 
-    public Atleta(double dni, String nombre, String apellido, int edad, String nacionalidad, String especialidad){
+    public Atleta(long dni, String nombre, String apellido, int edad, String nacionalidad, String especialidad){
         super(dni, nombre, apellido, edad, nacionalidad);
         this.especialidad = especialidad;
     }
